# oba-ui/src/app/api/swagger/oba.json
{
	"swagger": "2.0",
	"info": {
		"description": "Implementation of backend for online banking UI. We have 3 preloaded users in Ledgers: <b>marion.mueller</b>, <b>anton.brueckner</b>, <b>max.musterman</b> all with the PIN <b>12345</b>. You can use the User Login API from Ledgers <b>/users/authorise2</b> endpoint to gain an access token. Then use the access token with the prefix 'Bearer ' to Authorize on this ui.",
		"version": "0.5.0",
		"title": "Online banking",
		"termsOfService": "Terms of Service: to be edited...",
		"contact": {
			"name": "OFIN",
			"url": "https://www.ofin.co",
			"email": "fpo@ofin.co"
		},
		"license": {
			"name": "Apache License Version 2.0",
			"url": "https://www.apache.org/licenses/LICENSE-2.0"
		}
	},
	"host": "dev-dynamicsandbox-onlinebanking.cloud.ofin.co",
	"basePath": "/",
...
	"paths": {
		"/ais/accounts": {
			"get": {
				"tags": [
					"PSU AIS. Provides access to online banking account functionality"
				],
				"summary": "List fo Accessible Accounts",
				"description": "Returns the list of all accounts linked to the connected user. Call only available to role CUSTOMER.",
				"operationId": "getListOfAccountsUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "Cookie",
						"in": "header",
						"description": "Cookie",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "List of accounts accessible to the user.",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/AccountDetailsTO"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"deprecated": false
			}
		},
...
		"/ais/{encryptedConsentId}/authorisation/{authorisationId}/authCode": {
			"post": {
				"tags": [
					"PSU AIS. Provides access to online banking account functionality"
				],
				"summary": "Provides a TAN for the validation of an authorization",
				"operationId": "authrizedConsentUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "Cookie",
						"in": "header",
						"description": "Cookie",
						"required": false,
						"type": "string"
					},
					{
						"name": "authCode",
						"in": "query",
						"description": "authCode",
						"required": true,
						"type": "string"
					},
					{
						"name": "authorisationId",
						"in": "path",
						"description": "authorisationId",
						"required": true,
						"type": "string"
					},
					{
						"name": "encryptedConsentId",
						"in": "path",
						"description": "encryptedConsentId",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/ConsentAuthorizeResponse"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"deprecated": false
			}
		},
		"/ais/{encryptedConsentId}/authorisation/{authorisationId}/done": {
			"get": {
				"tags": [
					"PSU AIS. Provides access to online banking account functionality"
				],
				"summary": "Close consent session",
				"description": "This call provides the server with the opportunity to close this session and redirect the PSU to the TPP or close the application window.",
				"operationId": "aisDoneUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "Cookie",
						"in": "header",
						"description": "Cookie",
						"required": false,
						"type": "string"
					},
					{
						"name": "authConfirmationCode",
						"in": "query",
						"description": "authConfirmationCode",
						"required": false,
						"type": "string"
					},
					{
						"name": "authorisationId",
						"in": "path",
						"description": "authorisationId",
						"required": true,
						"type": "string"
					},
					{
						"name": "encryptedConsentId",
						"in": "path",
						"description": "encryptedConsentId",
						"required": true,
						"type": "string"
					},
					{
						"name": "oauth2",
						"in": "query",
						"description": "oauth2",
						"required": false,
						"type": "boolean",
						"default": false
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/ConsentAuthorizeResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"deprecated": false
			}
		},
...
		"/ais/{encryptedConsentId}/authorisation/{authorisationId}/methods/{scaMethodId}": {
			"post": {
				"tags": [
					"PSU AIS. Provides access to online banking account functionality"
				],
				"summary": "Selects the SCA Method for use.",
				"operationId": "selectMethodUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "Cookie",
						"in": "header",
						"description": "Cookie",
						"required": false,
						"type": "string"
					},
					{
						"name": "authorisationId",
						"in": "path",
						"description": "authorisationId",
						"required": true,
						"type": "string"
					},
					{
						"name": "encryptedConsentId",
						"in": "path",
						"description": "encryptedConsentId",
						"required": true,
						"type": "string"
					},
					{
						"name": "scaMethodId",
						"in": "path",
						"description": "scaMethodId",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/ConsentAuthorizeResponse"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"deprecated": false
			}
		},
		"/ais/{encryptedConsentId}/authorisation/{authorisationId}/start": {
			"post": {
				"tags": [
					"PSU AIS. Provides access to online banking account functionality"
				],
				"summary": "Starts the cosent authaurization process after user selects which account to grant access to",
				"operationId": "startConsentAuthUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "Cookie",
						"in": "header",
						"description": "Cookie",
						"required": false,
						"type": "string"
					},
					{
						"in": "body",
						"name": "aisConsent",
						"description": "aisConsent",
						"required": true,
						"schema": {
							"$ref": "#/definitions/AisConsentRequest"
						}
					},
					{
						"name": "authorisationId",
						"in": "path",
						"description": "authorisationId",
						"required": true,
						"type": "string"
					},
					{
						"name": "encryptedConsentId",
						"in": "path",
						"description": "encryptedConsentId",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/ConsentAuthorizeResponse"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"deprecated": false
			}
		},
		"/ais/{encryptedConsentId}/{authorisationId}": {
			"delete": {
				"tags": [
					"PSU AIS. Provides access to online banking account functionality"
				],
				"summary": "Revoke consent",
				"description": "This call provides the server with the opportunity to close this session and revoke consent.",
				"operationId": "revokeConsentUsingDELETE",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "Cookie",
						"in": "header",
						"description": "Cookie",
						"required": false,
						"type": "string"
					},
					{
						"name": "authorisationId",
						"in": "path",
						"description": "authorisationId",
						"required": true,
						"type": "string"
					},
					{
						"name": "encryptedConsentId",
						"in": "path",
						"description": "encryptedConsentId",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/ConsentAuthorizeResponse"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"deprecated": false
			}
		},
		"/api/v1/ais/account/{accountId}": {
			"get": {
				"tags": [
					"Online Banking Account Information"
				],
				"summary": "Get account details by account id",
				"operationId": "accountUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "accountId",
						"in": "path",
						"description": "accountId",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/AccountDetailsTO"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"deprecated": false
			}
		},
		"/api/v1/ais/accounts/{userLogin}": {
			"get": {
				"tags": [
					"Online Banking Account Information"
				],
				"summary": "Get List of users accounts",
				"operationId": "accountsUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "userLogin",
						"in": "path",
						"description": "userLogin",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/AccountDetailsTO"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"deprecated": false
			}
		},
...
		"/api/v1/ais/transactions/{accountId}": {
			"get": {
				"tags": [
					"Online Banking Account Information"
				],
				"summary": "Get List of transactions for queried account per dates selected",
				"operationId": "transactionsUsingGET_1",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "accountId",
						"in": "path",
						"description": "accountId",
						"required": true,
						"type": "string"
					},
					{
						"name": "dateFrom",
						"in": "query",
						"description": "dateFrom",
						"required": false,
						"type": "string",
						"format": "date"
					},
					{
						"name": "dateTo",
						"in": "query",
						"description": "dateTo",
						"required": false,
						"type": "string",
						"format": "date"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/TransactionTO"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"deprecated": false
			}
		},
		"/api/v1/ais/transactions/{accountId}/page": {
			"get": {
				"tags": [
					"Online Banking Account Information"
				],
				"summary": "Get List of transactions for queried account per dates selected, paged view",
				"operationId": "transactionsUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "accountId",
						"in": "path",
						"description": "accountId",
						"required": true,
						"type": "string"
					},
					{
						"name": "dateFrom",
						"in": "query",
						"description": "dateFrom",
						"required": false,
						"type": "string",
						"format": "date"
					},
					{
						"name": "dateTo",
						"in": "query",
						"description": "dateTo",
						"required": false,
						"type": "string",
						"format": "date"
					},
					{
						"name": "page",
						"in": "query",
						"description": "page",
						"required": false,
						"type": "integer",
						"default": 0,
						"format": "int32"
					},
					{
						"name": "size",
						"in": "query",
						"description": "size",
						"required": false,
						"type": "integer",
						"default": 25,
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/CustomPageImpl«TransactionTO»"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"deprecated": false
			}
		},
...
		"/api/v1/consents/{consentId}": {
			"put": {
				"tags": [
					"Online Banking Consents"
				],
				"summary": "Revoke consent by ID",
				"operationId": "revokeConsentUsingPUT",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "consentId",
						"in": "path",
						"description": "consentId",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "boolean"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"deprecated": false
			}
		},
...
		"/oauth/authorise": {
			"post": {
				"tags": [
					"Online Banking Oauth Authorization"
				],
				"summary": "Oauth authorise",
				"operationId": "oauthCodeUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "login",
						"in": "header",
						"description": "login",
						"required": true,
						"type": "string"
					},
					{
						"name": "pin",
						"in": "header",
						"description": "pin",
						"required": true,
						"type": "string"
					},
					{
						"name": "redirect_uri",
						"in": "query",
						"description": "redirect_uri",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/OauthCodeResponseTO"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"deprecated": false
			}
		},
...
		"/pis-cancellation/{encryptedPaymentId}/authorisation/{authorisationId}/methods/{scaMethodId}": {
			"post": {
				"tags": [
					"PSU PIS Cancellation. Provides access to online banking payment functionality"
				],
				"summary": "Selects the SCA Method for use.",
				"operationId": "selectMethodUsingPOST_1",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "Cookie",
						"in": "header",
						"description": "Cookie",
						"required": false,
						"type": "string"
					},
					{
						"name": "authorisationId",
						"in": "path",
						"description": "authorisationId",
						"required": true,
						"type": "string"
					},
					{
						"name": "encryptedPaymentId",
						"in": "path",
						"description": "encryptedPaymentId",
						"required": true,
						"type": "string"
					},
					{
						"name": "scaMethodId",
						"in": "path",
						"description": "scaMethodId",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/PaymentAuthorizeResponse"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"deprecated": false
			}
		},
...
		"/pis/{encryptedPaymentId}/authorisation/{authorisationId}/methods/{scaMethodId}": {
			"post": {
				"tags": [
					"PSU PIS. Provides access to online banking payment functionality"
				],
				"summary": "Selects the SCA Method for use.",
				"operationId": "selectMethodUsingPOST_2",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "Cookie",
						"in": "header",
						"description": "Cookie",
						"required": false,
						"type": "string"
					},
					{
						"name": "authorisationId",
						"in": "path",
						"description": "authorisationId",
						"required": true,
						"type": "string"
					},
					{
						"name": "encryptedPaymentId",
						"in": "path",
						"description": "encryptedPaymentId",
						"required": true,
						"type": "string"
					},
					{
						"name": "scaMethodId",
						"in": "path",
						"description": "scaMethodId",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/PaymentAuthorizeResponse"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"security": [
					{
						"apiKey": []
					}
				],
				"deprecated": false
			}
		},
...
	},
	"securityDefinitions": {
		"apiKey": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	},
	"definitions": {
...
		"AccountDetailsTO": {
			"type": "object",
			"properties": {
				"accountStatus": {
					"type": "string",
					"enum": [
						"ENABLED",
						"DELETED",
						"BLOCKED"
					]
				},
				"accountType": {
					"type": "string",
					"enum": [
						"CACC",
						"CASH",
						"CHAR",
						"CISH",
						"COMM",
						"CPAC",
						"LLSV",
						"LOAN",
						"MGLD",
						"MOMA",
						"NREX",
						"ODFT",
						"ONDP",
						"OTHR",
						"SACC",
						"SLRY",
						"SVGS",
						"TAXE",
						"TRAN",
						"TRAS"
					]
				},
				"balances": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/AccountBalanceTO"
					}
				},
				"bban": {
					"type": "string"
				},
				"bic": {
					"type": "string"
				},
				"currency": {
					"type": "string"
				},
				"details": {
					"type": "string"
				},
				"iban": {
					"type": "string"
				},
				"id": {
					"type": "string"
				},
				"linkedAccounts": {
					"type": "string"
				},
				"maskedPan": {
					"type": "string"
				},
				"msisdn": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"pan": {
					"type": "string"
				},
				"product": {
					"type": "string"
				},
				"usageType": {
					"type": "string",
					"enum": [
						"PRIV",
						"ORGA"
					]
				}
			},
			"title": "AccountDetailsTO"
		},
...
		"OauthCodeResponseTO": {
			"type": "object",
			"properties": {
				"redirectUri": {
					"type": "string"
				}
			},
			"title": "OauthCodeResponseTO"
		},
...
	}
}